Extract PinwheelChrome (the SwiftUI<->window seam) into its own file; keep the window-host plumbing together
