+++ Pinwheel/Sources/Pinwheel/Public/PinwheelFloatingControlsHost.swift
@@ -1,40 +1,6 @@
 import SwiftUI
 import UIKit
 
-/// Coordinates the floating tweak/close controls — the UIKit `CornerAnchoringView`
-/// hosted in a pass-through overlay window — with the SwiftUI catalog/preview.
-///
-/// The controls live in a separate `UIWindow` so they float *above* sheet
-/// presentations and never clip to (or clash with) the presented content. The
-/// SwiftUI side only reads/writes this coordinator; the window observes it.
-@MainActor
-@Observable
-final class PinwheelChrome {
-    /// The presented component's tweaks. Held here (a persistent reference) rather
-    /// than in the playground's `@State` so they survive playground re-renders /
-    /// identity changes — the deep-link preview otherwise lost them when the
-    /// settings sheet opened, leaving the sheet empty.
-    var tweaks: [PinwheelTweak] = []
-    /// Whether a component is currently presented — gates the FAB's visibility.
-    var isPresentingItem: Bool = false
-    /// Drives the SwiftUI settings sheet; the wrench button sets this `true`.
-    var showsSettings: Bool = false
-    /// Dismisses the presented component; the close button invokes this.
-    var onClose: (() -> Void)?
-
-    /// Badge count shown on the tweak (wrench) button.
-    var tweakCount: Int { tweaks.count }
-
-    /// The FAB shows only while a component is presented and the settings sheet
-    /// is closed (so it never floats over the settings sheet itself).
-    var isFloatingControlsVisible: Bool {
-        isPresentingItem && !showsSettings
-    }
-
-    func selectSettings() { showsSettings = true }
-    func selectClose() { onClose?() }
-}
-
 /// Installs a pass-through overlay `UIWindow` hosting the UIKit
 /// `CornerAnchoringView` FAB and pushes `PinwheelChrome` state into it. Placed as
 /// a background of the catalog/preview root; touches outside the FAB buttons fall
